+++ src/engine/Join.h
@@ -37,8 +37,6 @@ class Join : public Operation {
 
   using OptionalPermutation = std::optional<std::vector<ColumnIndex>>;
 
-  static constexpr size_t CHUNK_SIZE = 100'000;
-
   virtual string getDescriptor() const override;
 
   virtual size_t getResultWidth() const override;
@@ -85,21 +83,6 @@ class Join : public Operation {
    * a proper switch.
    **/
   void join(const IdTable& a, const IdTable& b, IdTable* result) const;
-
- private:
-  // Part of the implementation of `createResult`. This function is called when
-  // the result should be yielded lazily.
-  // Action is a lambda that itself runs the join operation in a blocking
-  // manner. It is passed a special function that is supposed to be the callback
-  // being passed to the `AddCombinedRowToIdTable` so that the partial results
-  // can be yielded during execution. This is achieved by spawning a separate
-  // thread.
-  CPP_template_2(typename ActionT)(
-      requires ad_utility::InvocableWithExactReturnType<
-          ActionT, Result::IdTableVocabPair,
-          std::function<void(IdTable&, LocalVocab&)>>) Result::Generator
-      runLazyJoinAndConvertToGenerator(ActionT action,
-                                       OptionalPermutation permutation) const;
 
  public:
   // Helper function to compute the result of a join operation and conditionally
