+++ test/JoinTest.cpp
@@ -20,6 +20,7 @@
 #include "engine/Engine.h"
 #include "engine/IndexScan.h"
 #include "engine/Join.h"
+#include "engine/JoinHelpers.h"
 #include "engine/OptionalJoin.h"
 #include "engine/QueryExecutionTree.h"
 #include "engine/Values.h"
@@ -34,6 +35,8 @@
 
 using ad_utility::testing::makeAllocator;
 namespace {
+
+using qlever::joinHelpers::CHUNK_SIZE;
 
 using Vars = std::vector<std::optional<Variable>>;
 auto iri = [](std::string_view s) {
@@ -208,13 +211,6 @@ std::vector<JoinTestCase> createJoinTestSet() {
                    makeIdTableFromVector(expectedResult), true});
 
   return myTestSet;
-}
-
-IdTable createIdTableOfSizeWithValue(size_t size, Id value) {
-  IdTable idTable{1, ad_utility::testing::makeAllocator()};
-  idTable.resize(size);
-  ql::ranges::fill(idTable.getColumn(0), value);
-  return idTable;
 }
 }  // namespace
 
@@ -674,8 +670,8 @@ TEST(JoinTest, joinTwoLazyOperationsWithAndWithoutUndefValues) {
 
   leftTables.push_back(makeIdTableFromVector({{U}}));
   leftTables.push_back(makeIdTableFromVector({{I(2)}}));
-  rightTables.push_back(createIdTableOfSizeWithValue(Join::CHUNK_SIZE, I(1)));
-  auto expected7 = createIdTableOfSizeWithValue(Join::CHUNK_SIZE, I(1));
+  rightTables.push_back(createIdTableOfSizeWithValue(CHUNK_SIZE, I(1)));
+  auto expected7 = createIdTableOfSizeWithValue(CHUNK_SIZE, I(1));
   performJoin(std::move(leftTables), std::move(rightTables), expected7, false);
 
   leftTables.push_back(makeIdTableFromVector({{U}}));
@@ -759,8 +755,8 @@ TEST(JoinTest, joinLazyAndNonLazyOperationWithAndWithoutUndefValues) {
 
   rightTables.push_back(makeIdTableFromVector({{U}}));
   rightTables.push_back(makeIdTableFromVector({{I(2)}}));
-  auto expected7 = createIdTableOfSizeWithValue(Join::CHUNK_SIZE, I(1));
-  performJoin(createIdTableOfSizeWithValue(Join::CHUNK_SIZE, I(1)),
+  auto expected7 = createIdTableOfSizeWithValue(CHUNK_SIZE, I(1));
+  performJoin(createIdTableOfSizeWithValue(CHUNK_SIZE, I(1)),
               std::move(rightTables), expected7, false);
 }
