+++ test/engine/CMakeLists.txt
@@ -14,3 +14,4 @@ addLinkAndDiscoverTestSerial(QueryExecutionTreeTest engine)
 addLinkAndDiscoverTestSerial(DescribeTest engine)
 addLinkAndDiscoverTestSerial(ExistsJoinTest engine)
 addLinkAndDiscoverTest(NeutralOptionalTest engine)
+addLinkAndDiscoverTest(OptionalJoinTest engine)
