+++ test/util/IdTableHelpers.cpp
@@ -260,3 +260,11 @@ std::pair<IdTable, std::vector<LocalVocab>> aggregateTables(
   }
   return {std::move(aggregateTable), std::move(localVocabs)};
 }
+
+// _____________________________________________________________________________
+IdTable createIdTableOfSizeWithValue(size_t size, Id value) {
+  IdTable idTable{1, ad_utility::testing::makeAllocator()};
+  idTable.resize(size);
+  ql::ranges::fill(idTable.getColumn(0), value);
+  return idTable;
+}
